--- plugins/script-node-step-plugin/src/main/groovy/org/rundeck/plugin/scriptnodestep/ScriptURLNodeStepPlugin.groovy
@@ -1,256 +1,0 @@
-package org.rundeck.plugin.scriptnodestep
-
-import com.dtolabs.rundeck.core.common.INodeEntry
-import com.dtolabs.rundeck.core.common.UpdateUtils
-import com.dtolabs.rundeck.core.common.impl.URLFileUpdater
-import com.dtolabs.rundeck.core.common.impl.URLFileUpdaterBuilder
-import com.dtolabs.rundeck.core.data.BaseDataContext
-import com.dtolabs.rundeck.core.data.DataContext
-import com.dtolabs.rundeck.core.data.MultiDataContext
-import com.dtolabs.rundeck.core.data.SharedDataContextUtils
-import com.dtolabs.rundeck.core.dispatcher.ContextView
-import com.dtolabs.rundeck.core.dispatcher.DataContextUtils
-import com.dtolabs.rundeck.core.execution.workflow.WFSharedContext
-import com.dtolabs.rundeck.core.execution.workflow.steps.FailureReason
-import com.dtolabs.rundeck.core.execution.workflow.steps.StepFailureReason
-import com.dtolabs.rundeck.core.execution.workflow.steps.node.NodeStepException
-import com.dtolabs.rundeck.core.execution.workflow.steps.node.impl.DefaultScriptFileNodeStepUtils
-import com.dtolabs.rundeck.core.execution.workflow.steps.node.impl.ScriptFileNodeStepUtils
-import com.dtolabs.rundeck.core.execution.workflow.steps.node.impl.ScriptURLNodeStepExecutor
-import com.dtolabs.rundeck.core.plugins.Plugin
-import com.dtolabs.rundeck.core.utils.Converter
-import com.dtolabs.rundeck.core.utils.OptsUtil
-import com.dtolabs.rundeck.plugins.ServiceNameConstants
-import com.dtolabs.rundeck.plugins.descriptions.PluginDescription
-import com.dtolabs.rundeck.plugins.descriptions.PluginProperty
-import com.dtolabs.rundeck.plugins.step.NodeStepPlugin
-import com.dtolabs.rundeck.plugins.step.PluginStepContext
-import org.apache.commons.codec.binary.Hex
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-
-import java.nio.charset.Charset
-import java.security.MessageDigest
-import java.security.NoSuchAlgorithmException
-
-@Plugin(service = ServiceNameConstants.WorkflowNodeStep, name = ScriptURLNodeStepPlugin.PROVIDER_NAME)
-@PluginDescription(title = "Script URL Node Step", description = "Verify and validate design")
-class ScriptURLNodeStepPlugin implements NodeStepPlugin  {
-    public static final Logger logger                      = LoggerFactory.getLogger(ScriptURLNodeStepPlugin.class.getName());
-    public static final String PROVIDER_NAME = "script-url-node-step-plugin";
-
-    @PluginProperty(title = "URL",
-            description = "Url",
-            required = true)
-    private String path;
-
-    @PluginProperty(title = "Arguments",
-            description = "",
-            required = false)
-    private String arguments;
-
-    @PluginProperty(title = "Invocation String",
-            description = "",
-            required = false)
-    private String invocationString;
-
-    @PluginProperty(title = "Expand variables in script file",
-            description = "",
-            required = false)
-    private boolean expandVars;
-
-    @PluginProperty(title = "Quote arguments to script invocation string?",
-            description = "",
-            required = false)
-    private boolean quoteArgs;
-
-    @PluginProperty(title = "File Extension",
-            description = "",
-            required = false)
-    private String fileExtension;
-
-    private static final int DEFAULT_TIMEOUT = 30;
-    private static final boolean USE_CACHE = true;
-    public static final String UTF_8 = "UTF-8";
-
-    private File cacheDir;
-
-    private ScriptFileNodeStepUtils scriptUtils = new DefaultScriptFileNodeStepUtils();
-
-    private static String hashURL(final String url) {
-        try {
-            MessageDigest digest = MessageDigest.getInstance("SHA-1");
-            digest.reset();
-            digest.update(url.getBytes(Charset.forName(UTF_8)));
-            return new String(Hex.encodeHex(digest.digest()));
-        } catch (NoSuchAlgorithmException e) {
-            e.printStackTrace();
-        }
-        return Integer.toString(url.hashCode());
-    }
-
-    static enum Reason implements FailureReason{
-        /**
-         * Failed to download required URL
-         */
-        URLDownloadFailure
-    }
-
-    public void executeNodeStep(PluginStepContext context, Map<String, Object> configuration, INodeEntry entry) throws NodeStepException {
-        File destinationTempFile = downloadURLToTempFile(context, entry);
-        if (!USE_CACHE) {
-            destinationTempFile.deleteOnExit();
-        }
-        boolean expandTokens = true;
-        if (context.getFramework().hasProperty("execution.script.tokenexpansion.enabled")) {
-            expandTokens = "true".equals(context.getFramework().getProperty("execution.script.tokenexpansion.enabled"));
-        }
-
-        final String[] args;
-        if (null != arguments) {
-            args = OptsUtil.burst(arguments);
-        } else {
-            args = new String[0];
-        }
-
-        scriptUtils.executeScriptFile(
-                context,
-                entry,
-                null,
-                destinationTempFile.getAbsolutePath(),
-                null,
-                fileExtension,
-                args,
-                invocationString,
-                quoteArgs,
-                context.getFramework().getStepExecutionService(),
-                expandTokens
-        );
-    }
-
-    private File downloadURLToTempFile(
-            final PluginStepContext context,
-            final INodeEntry node
-    ) throws NodeStepException
-    {
-        if (!cacheDir.isDirectory() && !cacheDir.mkdirs()) {
-            throw new RuntimeException("Unable to create cachedir: " + cacheDir.getAbsolutePath());
-        }
-        //create node context for node and substitute data references in command
-        WFSharedContext sharedContext = new WFSharedContext(context.getSharedDataContext());
-        sharedContext.merge(ContextView.global(), context.getDataContextObject());
-        sharedContext.merge(
-                ContextView.node(node.getNodename()),
-                new BaseDataContext("node", DataContextUtils.nodeData(node))
-        );
-
-
-        final String finalUrl = expandUrlString(path, sharedContext, node.getNodename());
-        final URL url;
-        try {
-            url = new URL(finalUrl);
-        } catch (MalformedURLException e) {
-            throw new NodeStepException(e, StepFailureReason.ConfigurationFailure, node.getNodename());
-        }
-        if(null!=context.getExecutionListener()){
-            context.getExecutionListener().log(4, "Requesting URL: " + url.toExternalForm());
-        }
-
-        String cleanUrl = url.toExternalForm().replaceAll("^(https?://)([^:@/]+):[^@/]*@", "\$1\$2:****@");
-        String tempFileName = hashURL(url.toExternalForm()) + ".temp";
-        File destinationTempFile = new File(cacheDir, tempFileName);
-        File destinationCacheData = new File(cacheDir, tempFileName + ".cache.properties");
-
-        //update from URL if necessary
-        final URLFileUpdaterBuilder urlFileUpdaterBuilder = new URLFileUpdaterBuilder()
-                .setUrl(url)
-                .setAcceptHeader("*/*")
-                .setTimeout(DEFAULT_TIMEOUT);
-        if (USE_CACHE) {
-            urlFileUpdaterBuilder
-                    .setCacheMetadataFile(destinationCacheData)
-                    .setCachedContent(destinationTempFile)
-                    .setUseCaching(true);
-        }
-        final URLFileUpdater updater = urlFileUpdaterBuilder.createURLFileUpdater();
-        try {
-            UpdateUtils.update(updater, destinationTempFile);
-
-            logger.debug("Updated nodes resources file: " + destinationTempFile);
-        } catch (UpdateUtils.UpdateException e) {
-            if (!destinationTempFile.isFile() || destinationTempFile.length() < 1) {
-                throw new NodeStepException(
-                        "Error requesting URL Script: " + cleanUrl + ": " + e.getMessage(),
-                        e,
-                        ScriptURLNodeStepExecutor.Reason.URLDownloadFailure,
-                        node.getNodename());
-            } else {
-                logger.error(
-                        "Error requesting URL script: " + cleanUrl + ": " + e.getMessage(), e);
-            }
-        }
-        return destinationTempFile;
-    }
-
-    public static final Converter<String, String> urlPathEncoder = s -> {
-        try {
-            return URLEncoder.encode(s,UTF_8).replace("+","%20");
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-            return s;
-        }
-    };
-    public static final Converter<String, String> urlQueryEncoder = s -> {
-        try {
-            return URLEncoder.encode(s,UTF_8).replace("+","%20");
-        } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
-            return s;
-        }
-    };
-
-    /**
-     * Expand data references in a URL string, using proper encoding for path and query parts.
-     * @param urlString url
-     * @param dataContext multi context
-     * @param nodename default node context
-     * @return expanded string
-     */
-    public static String expandUrlString(
-            final String urlString,
-            final MultiDataContext<ContextView, DataContext> dataContext,
-            final String nodename
-    )
-    {
-        final int qindex = urlString.indexOf('?');
-        final StringBuilder builder = new StringBuilder();
-
-        builder.append(
-                SharedDataContextUtils.replaceDataReferences(
-                        qindex > 0 ? urlString.substring(0, qindex) : urlString,
-                        dataContext,
-                        ContextView.node(nodename),
-                        ContextView::nodeStep,
-                        DataContextUtils.replaceMissingOptionsWithBlank,
-                        true,
-                        false
-                ));
-        if (qindex > 0) {
-            builder.append("?");
-            if (qindex < urlString.length() - 1) {
-                builder.append(SharedDataContextUtils.replaceDataReferences(
-                        urlString.substring(qindex + 1),
-                        dataContext,
-                        ContextView.node(nodename),
-                        ContextView::nodeStep,
-                        urlQueryEncoder,
-                        true,
-                        false
-                ));
-            }
-
-        }
-
-        return builder.toString();
-    }
-}
